+++ RELEASES.md
@@ -1,3 +1,10 @@
+# Release 0.2.3 (2018-10-18)
+
+* `seq` function - Same as R
+* Extract Col & Row
+    * `a.col(1)` : Extract 1st column of `a` as Column matrix
+    * `a.row(1)` : Extract 1st row of `a` as Row matrix
+
 # Release 0.2.1 (2018-10-04)
 
 * Update Documentation